+++ src/decorators/BuiltInstance.ts
@@ -217,7 +217,7 @@ export class BuiltInstance {
             loggerFactory
                 .getLogger()
                 .warn(
-                    '[mnt] old build?',
+                    '[mtn] old build?',
                     old,
                     build.id,
                     delta,